Add header definitions for markers
There's one main marker header, and then an individual header for each
marker item.

There's a fair amount of difference between the specced definitions in
3.x and the ones we're seeing in 4.2, but they're similar enough to 
be useful.

This is trying a different technique than the crazy if/then switching
for __h_elts we've used in the past.

Really, all of this should die in a fire and be replaced by ctypes.