Switch struct_dict to use OrderedDicts
This will make debugging easier -- the data will stay in the original
order.

The *right* answer here would be to get rid of StructDict in favor of
ctypes, but that project would be a clusterfuck for identical
functionality. Anyone else wants to take it on, there's a good test
suite now.