Always write matlab data in system byte order
Our data arrays can be in different byte orders. For example, the .acq
data can contain a mix of floats and ints in a byte ordering that
doesn't match the system's ordering. In that case, the float channels
will be in the .acq file's ordering, and the float-ified int channels
will be in the system's native ordering. Chaos ensues.

In this case, making sure our channels are in the system's native byte
order (dtype '=f8') will give us a working .mat file.