Try to hack in better foreign header support
This one is tricky -- the length for the header is normally 8, but in
this case it's much larger (3814) and the channel dtype headers *don't*
start immediately afterwards, but 40 bytes after that.

This adds a hack where we take the length, and if it's larger than 8, we
use the length plus 40. I don't know if this will work in the general
case.