Guess file encoding from version
It looks like "old" files are encoded in latin1, while "new" files are
encoded in utf8. I'm guessing that this switch happened at acq4, since
a lot of things happened there.

For now, at least, this removes the ability to specify an encoding when
reading these files. If there's a reason to re-add it, we can do that
later; for now, I think the guess is probably the right way to go.

The main place we'd see this is in old files that have µV units. I've
only seen a couple examples of this, so it may not be a complete fix.

For added robustness, I've added errors='ignore' to all decode() calls,
so we'll drop characters we can't identify instead of crashing. The
downside to this is that we may miss future text encoding bugs; who's
going to report a "weird marker text" bug? But ultimately we probably
care more about the data being read than the metadata being interpreted
correctly.

This probably fixes #27.