Scan for channel data type headers
In some files with non-empty foreign data headers, the length field of
the header isn't sufficient to seek to the channel data type headers.
That's weird! I can't find anything that clues us in to how far we need
to look ahead.

However! We have one saving grace here: The channel dtype headers have a
specific-enough format that we can potentially just scan forward from
the end of the foreign data header to look for them. We try to read the
headers, see if they're all valid, and if not, walk forward one byte.
Right now, I'm allowing 4096 tries -- the two files I have where the
foreign data header's length is wrong are off by 36 and 80 bytes, so
this will hopefully be more than enough.

There's a possibility that this is wrong, and that it'll read things
wrong and _not_ crash -- however, we should crash spectacularly when we
get to the markers and journal, so I'm not too worried that we'll
"successfully" decode incorrect data.