Fix all known journal reading problems
Thanks to the work (many years ago now!) from friend of the show Mike
Davison, this commit revamps the journal reading code for pre-V4 files.
Now it correctly finds and reads a set of marker metadata headers
instead of just saying "well I don't know what this block of stuff is"
and letting Jesus take the wheel to find the journal, this should locate
it a lot more precisely.