Switch headers to use ctypes
This was probably a silly thing to do, but it was fun.

This commit switches out the tried-and-true, but really silly,
StructDict thing I had made, in favor of an implementation with ctypes
Structures.

This should(!!!!) be effectively identical in behavior -- other than
some formatting changes to debugging messages, the data should come out
the same as before.

This does want a bit of testing before wide release.