Refactor tests, fix bugs
Now, we're really testing every single file in the test dataset. All the
files successfully read without error.

The only problem in the test dataset is one of the 5.0.1 files has
messed up marker sample indexes, which I genuinely don't understand. If
I get a chance, I'll try to explore that with a real copy of
AcqKnowledge.