Add acq_layout CLI tool
This adds a new utility, acq_layout, that prints the binary layout of an
.acq file in TSV format. This should make it easier to debug future
iterations of file formats.

This also refactors the readers one more time -- it removes a bunch of
unused code and weird circular dependencies and generally just makes
reading the headers and data a much more sequential process.
Reader._read_header() is still a little long, but there's not much logic
in it now.