Improvements to acq_reader
Now can (and usually does) truncate long arrays and data
representations.

Also shows a bit more information about the data in the file -- in
addition to start offsets and lengths, it reports the correct endianness
of the data.

For uncompressed data, it also reports the sample pattern, both in
"logical" (eg, which channels were collected in which order) and in
"data" layout (eg, which bytes in the data stream belong to which
channel) and includes a small sample of the data byte stream.

For compressed data, it reports the channel name, samples of compressed
and uncompressed data, and also the compression ratio.